RF: Put device exceptions in their own module
